chore: query를 위한 함수 분리 외
