+++ client/src/pages/user/update/UserImgFileUpdate.tsx
@@ -6,9 +6,10 @@ import FilePlusLabel from '../../../components/FilePlusLabel';
 import { useImgCrop } from '../../../hooks/useImgCrop';
 import { handleFileChange } from '../../../components/img-crop/imgCropUtils';
 import { store } from '../../../components/redux/store';
+import { UserInfoResData } from '../../../types/user';
 
 interface RandomAvatarUpdateProps {
-  userData: any | null; // 이 데이터 타입은 추후 api 작성 완료 되면 수정 예정
+  userData: UserInfoResData;
   setIsOpen: (arg: boolean) => void;
   isOpen: boolean;
 }
